layout/base/tests/test_bug849219.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=849219
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 849219</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 849219 **/

SimpleTest.waitForExplicitFinish();

function runTest() {
  var win = e.contentWindow;
  if (win.location.hash != '') {
    is(win.scrollY,0);
    SimpleTest.finish();
    return;
  }
  win.location.hash='#anchor'
  win.scrollTo(0,0);
  win.location.reload()
}


  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=849219">Mozilla Bug 849219</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
<script>

var e = document.createElement('iframe');
var url = 'file_bug842853-frame.html';
e.setAttribute('src',url);
e.setAttribute('onload','runTest()');
document.body.appendChild(e);

</script>
</body>
</html>
